[MISC] move seqan3::views::kmer_hash to seqan3/search/views/
