+++ test/include/seqan3/test/performance/naive_minimiser_hash.hpp
@@ -20,8 +20,8 @@
 #include <seqan3/core/range/type_traits.hpp>
 #include <seqan3/range/views/complement.hpp>
 #include <seqan3/range/views/detail.hpp>
-#include <seqan3/range/views/kmer_hash.hpp>
 #include <seqan3/range/views/zip.hpp>
+#include <seqan3/search/views/kmer_hash.hpp>
 #include <seqan3/search/views/minimiser.hpp>
 
 namespace seqan3::detail
