+++ test/performance/search/views/view_kmer_hash_benchmark.cpp
@@ -8,7 +8,7 @@
 #include <benchmark/benchmark.h>
 
 #include <seqan3/alphabet/nucleotide/dna4.hpp>
-#include <seqan3/range/views/kmer_hash.hpp>
+#include <seqan3/search/views/kmer_hash.hpp>
 #include <seqan3/test/performance/sequence_generator.hpp>
 #include <seqan3/test/performance/naive_kmer_hash.hpp>
 #include <seqan3/test/performance/units.hpp>
