+++ test/snippet/search/dream_index/interleaved_bloom_filter_clear.cpp
@@ -1,7 +1,7 @@
 #include <seqan3/alphabet/nucleotide/dna4.hpp>
 #include <seqan3/core/debug_stream.hpp>
-#include <seqan3/range/views/kmer_hash.hpp>
 #include <seqan3/search/dream_index/interleaved_bloom_filter.hpp>
+#include <seqan3/search/views/kmer_hash.hpp>
 
 using seqan3::operator""_dna4;
