+++ test/unit/range/views/CMakeLists.txt
@@ -30,7 +30,6 @@ seqan3_test(translate_test.cpp)
 seqan3_test(trim_quality_test.cpp)
 seqan3_test(single_pass_input_test.cpp)
 seqan3_test(get_test.cpp)
-seqan3_test(kmer_hash_test.cpp)
 seqan3_test(interleave_test.cpp)
 seqan3_test(minimiser_hash_test.cpp)
 seqan3_test(zip_test.cpp)
