+++ test/unit/search/views/CMakeLists.txt
@@ -1,1 +1,2 @@
-seqan3_test(minimiser_test.cpp)
+seqan3_test (kmer_hash_test.cpp)
+seqan3_test (minimiser_test.cpp)
