+++ test/unit/search/views/minimiser_test.cpp
@@ -14,8 +14,8 @@
 #include <seqan3/range/container/bitcompressed_vector.hpp>
 #include <seqan3/range/views/complement.hpp>
 #include <seqan3/range/views/drop.hpp>
-#include <seqan3/range/views/kmer_hash.hpp>
 #include <seqan3/range/views/take_until.hpp>
+#include <seqan3/search/views/kmer_hash.hpp>
 #include <seqan3/search/views/minimiser.hpp>
 #include <seqan3/test/expect_range_eq.hpp>
